Major Updates
- IP/Port settings added to config.opt
- IP/Port settings now configurable in Launcher Settings menu
- Client.Functions.OPT.cs GetInt(string name) method added
- IP/Port changed from readonly to normal
- Updated TODO notes
- Client.GUI.cs added assignIP() method
- Client.GUI.cs Cleaned up error handling when canceling a login attempt
- Client.GUI.cs now requests a graceful disconnection from server on
close attempt
- Reimplemented showing the Rappelz Client settings menu
- Client.Network.ServerManager.cs changed short port argument to int
- Client.Network.ServerPackets.cs added client disconnection packets
- Added several option set prcoedures (not all checkboxes work yet)
- Server.Functions.OTP.cs removed SetOTP() method [this is handled in
Server.Functions.UserHandler.cs now
- Server.Functions.OTP.cs changed all referenced to otp_end column with
expiration for clarity
- Server.Functions.UpdateHandler.cs removed OnUserDisconnecti() method
[this is now handled by UserHandler]
- Server.Network.ClientManager.cs added better logging of disconnection
+ client cleanup.
- Server.Network.ClientPackets.cs added client request packet
- Server.Portal.opt reformatted and future settings for multiple file
delivery system added

+++ Client/Network/ServerManager.cs
@@ -30,7 +30,7 @@ namespace Client.Network
         /// <param name="ip"></param>
         /// <param name="port"></param>
         /// <returns></returns>
-        public bool Start(string ip, short port)
+        public bool Start(string ip, int port)
         {
             Socket socket = new Socket(SocketType.Stream, ProtocolType.Tcp);
